+++ external-crates/move/crates/move-package-alt/src/errors/mod.rs
@@ -22,8 +22,8 @@ use thiserror::Error;
 use crate::dependency::ResolverError;
 use crate::git::GitError;
 use crate::graph::LinkageError;
-use crate::package::manifest::ManifestError;
 use crate::package::paths::PackagePathError;
+use crate::schema::ManifestError;
 
 /// Result type for package operations
 pub type PackageResult<T> = Result<T, PackageError>;
