+++ external-crates/move/crates/move-package-alt/src/schema/manifest.rs
@@ -1,17 +1,28 @@
-use std::{collections::BTreeMap, path::PathBuf};
+use std::{
+    collections::BTreeMap,
+    path::{Path, PathBuf},
+};
 
+use codespan_reporting::diagnostic::{Diagnostic, Label};
 use serde::{Deserialize, Deserializer, de};
 use serde_spanned::Spanned;
+use sha2::{Digest as ShaDigest, Sha256};
+use thiserror::Error;
 
 use super::{
     EnvironmentName, LocalDepInfo, OnChainDepInfo, PackageName, PublishAddresses, ResolverName,
 };
 
-// TODO: look at Brandon's serialization code (https://github.com/MystenLabs/sui-rust-sdk/blob/master/crates/sui-sdk-types/src/object.rs)
+use crate::errors::{FileHandle, Location};
+
+const ALLOWED_EDITIONS: &[&str] = &["2025", "2024", "2024.beta", "legacy"];
 
 /// The on-chain identifier for an environment (such as a chain ID); these are bound to environment
 /// names in the `[environments]` table of the manifest
 pub type EnvironmentID = String;
+
+pub type ManifestResult<T> = Result<T, ManifestError>;
+pub type Digest = String;
 
 // Note: [Manifest] objects are immutable and should not implement [serde::Serialize]; any tool
 // writing these files should use [toml_edit] to set / preserve the formatting, since these are
@@ -38,6 +49,8 @@ pub struct ParsedManifest {
 pub struct PackageMetadata {
     pub name: Spanned<PackageName>,
     pub edition: String,
+    #[serde(skip)]
+    pub digest: Digest,
 }
 
 /// An entry in the `[dependencies]` section of a manifest
@@ -116,6 +129,127 @@ struct RField {
     r: BTreeMap<String, toml::Value>,
 }
 
+#[derive(Error, Debug)]
+#[error("{kind}")]
+pub struct ManifestError {
+    pub kind: Box<ManifestErrorKind>,
+    location: ErrorLocation,
+}
+
+#[derive(Debug)]
+enum ErrorLocation {
+    WholeFile(PathBuf),
+    AtLoc(Location),
+}
+
+#[derive(Error, Debug)]
+pub enum ManifestErrorKind {
+    #[error("package name cannot be empty")]
+    EmptyPackageName,
+    #[error("unsupported edition '{edition}', expected one of '{valid}'")]
+    InvalidEdition { edition: String, valid: String },
+    #[error("externally resolved dependencies must have exactly one resolver field")]
+    BadExternalDependency,
+    #[error(
+        "dep-replacements.mainnet is invalid because mainnet is not in the [environments] table"
+    )]
+    MissingEnvironment { env: EnvironmentName },
+    #[error(
+        // TODO: add a suggested environment (needs to be part of the flavor)
+        "you must define at least one environment in the [environments] section of `Move.toml`."
+    )]
+    NoEnvironments,
+    #[error("{}", .0.message())]
+    ParseError(#[from] toml_edit::de::Error),
+    #[error(transparent)]
+    IoError(#[from] std::io::Error),
+}
+
+impl ParsedManifest {
+    pub fn read_from_file(file_handle: FileHandle) -> ManifestResult<Self> {
+        let mut parsed: ParsedManifest = toml_edit::de::from_str(file_handle.source())
+            .map_err(ManifestError::from_toml(file_handle))?;
+        parsed.set_digest(&file_handle);
+
+        parsed.validate_manifest(file_handle)?;
+
+        Ok(parsed)
+    }
+
+    /// Set the digest of the manifest to a SHA256 hash of the file contents.
+    fn set_digest(&mut self, file_id: &FileHandle) {
+        self.package.digest = format!("{:X}", Sha256::digest(file_id.source().as_ref()));
+    }
+
+    /// The combined entries of the `[dependencies]` and `[dep-replacements]` sections for this
+    /// manifest
+    pub fn dependencies(&self) -> BTreeMap<PackageName, DefaultDependency> {
+        self.dependencies
+            .iter()
+            .map(|(name, dep)| (name.as_ref().clone(), dep.clone()))
+            .collect()
+    }
+
+    pub fn dep_replacements(
+        &self,
+    ) -> &BTreeMap<EnvironmentName, BTreeMap<PackageName, Spanned<ReplacementDependency>>> {
+        &self.dep_replacements
+    }
+
+    pub fn metadata(&self) -> PackageMetadata {
+        self.package.clone()
+    }
+
+    /// The entries from the `[environments]` section
+    pub fn environments(&self) -> BTreeMap<EnvironmentName, EnvironmentID> {
+        self.environments
+            .iter()
+            .map(|(name, id)| (name.as_ref().clone(), id.as_ref().clone()))
+            .collect()
+    }
+
+    /// The name declared in the `[package]` section
+    pub fn package_name(&self) -> &PackageName {
+        self.package.name.as_ref()
+    }
+
+    /// A digest of the file, suitable for detecting changes
+    pub fn digest(&self) -> &Digest {
+        &self.package.digest
+    }
+
+    /// Validate the manifest contents, after deserialization.
+    ///
+    // TODO: add more validation
+    fn validate_manifest(&self, handle: FileHandle) -> ManifestResult<()> {
+        // Are there any environments?
+        if self.environments().is_empty() {
+            return Err(ManifestError::with_file(handle.path())(
+                ManifestErrorKind::NoEnvironments,
+            ));
+        }
+
+        // Do all dep-replacements have valid environments?
+        for (env, entries) in self.dep_replacements.iter() {
+            if !self.environments().contains_key(env) {
+                let span = entries
+                    .first_key_value()
+                    .expect("dep-replacements.<env> only exists if it has a dep")
+                    .1
+                    .span();
+
+                let loc = Location::new(handle, span);
+
+                return Err(ManifestError::with_span(&loc)(
+                    ManifestErrorKind::MissingEnvironment { env: env.clone() },
+                ));
+            }
+        }
+
+        Ok(())
+    }
+}
+
 impl<'de> Deserialize<'de> for ManifestDependencyInfo {
     fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
     where
@@ -164,5 +298,49 @@ impl TryFrom<RField> for ExternalDependency {
             .expect("iterator of length 1 structure is nonempty");
 
         Ok(Self { resolver, data })
+    }
+}
+
+impl ManifestError {
+    pub(crate) fn with_file<T: Into<ManifestErrorKind>>(
+        path: impl AsRef<Path>,
+    ) -> impl Fn(T) -> Self {
+        move |e| ManifestError {
+            kind: Box::new(e.into()),
+            location: ErrorLocation::WholeFile(path.as_ref().to_path_buf()),
+        }
+    }
+
+    fn with_span<T: Into<ManifestErrorKind>>(loc: &Location) -> impl Fn(T) -> Self {
+        move |e| ManifestError {
+            kind: Box::new(e.into()),
+            location: ErrorLocation::AtLoc(loc.clone()),
+        }
+    }
+
+    fn from_toml(file: FileHandle) -> impl Fn(toml_edit::de::Error) -> Self {
+        move |e| {
+            let location = e
+                .span()
+                .map(|span| ErrorLocation::AtLoc(Location::new(file, span)))
+                .unwrap_or(ErrorLocation::WholeFile(file.path().to_path_buf()));
+            ManifestError {
+                kind: Box::new(e.into()),
+                location,
+            }
+        }
+    }
+
+    /// Convert this error into a codespan Diagnostic
+    pub fn to_diagnostic(&self) -> Diagnostic<FileHandle> {
+        match &self.location {
+            ErrorLocation::WholeFile(path) => {
+                Diagnostic::error().with_message(format!("Error while loading `{path:?}`: {self}"))
+            }
+            ErrorLocation::AtLoc(loc) => Diagnostic::error()
+                .with_message(format!("Error while loading `{:?}`", loc.file()))
+                .with_labels(vec![Label::primary(loc.file(), loc.span().clone())])
+                .with_notes(vec![self.to_string()]),
+        }
     }
 }
