+++ external-crates/move/crates/move-package-alt/tests/test_runner.rs
@@ -8,9 +8,10 @@ use move_command_line_common::testing::insta_assert;
 use codespan_reporting::term::{self, Config, termcolor::Buffer};
 use move_package_alt::{
     dependency::{self, CombinedDependency, DependencySet, PinnedDependencyInfo},
-    errors::Files,
+    errors::{FileHandle, Files},
     flavor::{Vanilla, vanilla},
-    package::{RootPackage, lockfile::Lockfiles, manifest::Manifest, paths::PackagePath},
+    package::{RootPackage, lockfile::Lockfiles, paths::PackagePath},
+    schema::ParsedManifest,
 };
 use std::path::Path;
 use tracing::debug;
@@ -69,7 +70,8 @@ impl Test<'_> {
     fn output(&self) -> anyhow::Result<String> {
         Ok(match self.kind {
             "parsed" => {
-                let manifest = Manifest::<Vanilla>::read_from_file(self.toml_path);
+                let file_handle = FileHandle::new(self.toml_path).unwrap();
+                let manifest = ParsedManifest::read_from_file(file_handle);
                 let contents = match manifest.as_ref() {
                     Ok(m) => format!("{:#?}", m),
                     Err(_) => {
@@ -130,22 +132,21 @@ fn run_graph_to_lockfile_test_wrapper(path: &Path) -> Result<String, Box<dyn std
 }
 
 async fn run_pinning_tests(input_path: &Path) -> datatest_stable::Result<String> {
-    let manifest = Manifest::<Vanilla>::read_from_file(input_path).unwrap();
+    let file_handle = FileHandle::new(input_path).unwrap();
+    let env = vanilla::default_environment();
+    let manifest = ParsedManifest::read_from_file(file_handle).unwrap();
 
-    let deps: DependencySet<CombinedDependency> = manifest.dependencies();
+    let deps = CombinedDependency::combine_deps(file_handle, &env, &manifest).unwrap();
     let mut output = DependencySet::<PinnedDependencyInfo>::new();
     debug!("{deps:?}");
 
     add_bindir();
 
-    for env in manifest.environments().keys() {
-        let deps = deps.deps_for(env).unwrap();
-        let pinned = dependency::pin::<Vanilla>(deps.clone(), env)
-            .await
-            .map_err(|e| e.to_string())?;
-        for (name, dep) in pinned {
-            output.insert(env.clone(), name, dep);
-        }
+    let pinned = dependency::pin::<Vanilla>(deps.clone(), env.id())
+        .await
+        .map_err(|e| e.to_string())?;
+    for (name, dep) in pinned {
+        output.insert(env.name().to_string(), name, dep);
     }
 
     Ok(format!("{output:#?}"))
